fix(davinci): split scoped SFC tests under the source-length budget #4742
Moving the scoped-style regressions out of compile/tests.rs keeps that
already over-limit file from growing, which source-file-lengths rejects.
